Check EDITOR/OWNER 관련 PR

+++ src/main/kotlin/org/scaleadvisor/backend/project/application/service/projectlanguage/DeleteProjectLanguageService.kt
@@ -1,6 +1,8 @@
 package org.scaleadvisor.backend.project.application.service.projectlanguage
 
+import org.scaleadvisor.backend.global.exception.model.ForbiddenException
 import org.scaleadvisor.backend.project.application.port.repository.projectlanguage.DeleteProjectLanguagePort
+import org.scaleadvisor.backend.project.application.port.usecase.member.CheckIsEditorUseCase
 import org.scaleadvisor.backend.project.application.port.usecase.projectlanguage.DeleteProjectLanguageUseCase
 import org.scaleadvisor.backend.project.domain.id.ProjectId
 import org.springframework.stereotype.Service
@@ -9,9 +11,16 @@ import org.springframework.transaction.annotation.Transactional
 @Service
 @Transactional
 private class DeleteProjectLanguageService(
+    private val checkIsEditorUseCase: CheckIsEditorUseCase,
     private val deleteProjectLanguagePort: DeleteProjectLanguagePort
 ) : DeleteProjectLanguageUseCase {
     override fun deleteAll(projectId: ProjectId) {
+
+        if (!checkIsEditorUseCase.checkIsEditor(
+                projectId.toLong())) {
+            throw ForbiddenException("프로젝트 요소를 추가할 권한이 없습니다.")
+        }
+
         deleteProjectLanguagePort.deleteAll(projectId)
     }
 }
